requirements.txt
# AfriPlan Electrical v6.1 dependencies

# UI
streamlit>=1.30.0

# Data + validation
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# PDF pipeline
anthropic>=0.40.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# DXF pipeline
ezdxf>=1.4.0

# Exporters
openpyxl>=3.1.0
fpdf2>=2.7.0

# Plots / charts (optional, for any future analytics views)
matplotlib>=3.8.0
plotly>=5.18.0
